% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existence.R
\name{check_existence}
\alias{check_existence}
\title{Check the existence of the supplied address}
\usage{
check_existence(address)
}
\arguments{
\item{address}{Character vector of locations to check}
}
\value{
Invisible. Error if address does not exist
}
\description{
Check the existence of the supplied address
}
