% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_species_list.R
\name{collect_species_list}
\alias{collect_species_list}
\title{Query GBIF for a list of bird species found in a location}
\usage{
collect_species_list(location)
}
\arguments{
\item{location}{Character vector of location names passed to tidygeocoder or sf_POLYGON}
}
\value{
List of bird species recorded in GBIF within the area of interest
}
\description{
Query GBIF for a list of bird species found in a location
}
