% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recordings.R
\name{download_recordings}
\alias{download_recordings}
\title{Download recordings from multiple species}
\usage{
download_recordings(x, save_directory, n_per_sp = 1)
}
\arguments{
\item{x}{list of responses from search_xeno_canto()}

\item{save_directory}{directory to save the recordings}

\item{n_per_sp}{number of recordings to select from each species}
}
\value{
list of xeno canto records
}
\description{
Download recordings from multiple species
}
