% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recordings.R
\name{download_xeno}
\alias{download_xeno}
\title{Download a single Xeno Canto recording and image}
\usage{
download_xeno(x, save_directory)
}
\arguments{
\item{x}{a Xeno Canto record from search_xeno_canto()}

\item{save_directory}{directory to save the recordings}
}
\value{
local path to the recording
}
\description{
Download a single Xeno Canto recording and image
}
