% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recordings.R
\name{download_xeno_recording}
\alias{download_xeno_recording}
\title{Download a single Xeno Canto recording}
\usage{
download_xeno_recording(x_id, x_file, save_directory)
}
\arguments{
\item{x_id}{a Xeno Canto id number}

\item{x_file}{the url of a Xeno Canto file}

\item{save_directory}{the directory to save it in}
}
\value{
the local location of the file
}
\description{
Download a single Xeno Canto recording
}
