% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recordings.R
\name{download_xeno_sonogram}
\alias{download_xeno_sonogram}
\title{Download a single Xeno Canto sonogram}
\usage{
download_xeno_sonogram(x_id, x_sono_col, save_directory)
}
\arguments{
\item{x_id}{a Xeno Canto id number}

\item{x_sono_col}{the url of a Xeno Canto sonogram}

\item{save_directory}{the directory to save it in}
}
\value{
the local location of the sonogram
}
\description{
Download a single Xeno Canto sonogram
}
