% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recordings.R
\name{filter_recordings}
\alias{filter_recordings}
\title{Filter the list of recordings for a species to a given number}
\usage{
filter_recordings(x, n_per_sp = 1)
}
\arguments{
\item{x}{species record from Xeno Canto}

\item{n_per_sp}{number of recordings to keep from each species}
}
\value{
recording records
}
\description{
Filter the list of recordings for a species to a given number
}
