% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_geometry.R
\name{find_geometry}
\alias{find_geometry}
\title{Geocode and buffer a location}
\usage{
find_geometry(address, buffer_radius = 1e+05)
}
\arguments{
\item{address}{Character vector of locations to geocode}

\item{buffer_radius}{Numeric radius to buffer location in metres (default is 100000)}
}
\value{
sfc_POLYGON of buffered locations
}
\description{
Geocode and buffer a location
}
\examples{
find_geometry("Christmas Island, Australia")
find_geometry("France", buffer_radius = 1000000)
}
