% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_me.R
\name{flashcard}
\alias{flashcard}
\title{Test type: flashcard}
\usage{
flashcard(secret_record)
}
\arguments{
\item{secret_record}{a row from records_df}
}
\value{
invisible
}
\description{
Test type: flashcard
}
\examples{
\dontrun{flashcard(records_df[1])}
}
