% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_me.R
\name{free_choice}
\alias{free_choice}
\title{Test type: free choice}
\usage{
free_choice(secret_record)
}
\arguments{
\item{secret_record}{a row from records_df}
}
\value{
invisible
}
\description{
Test type: free choice
}
\examples{
\dontrun{free_choice(records_df[1])}
}
