% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recordings.R
\name{get_sono_large}
\alias{get_sono_large}
\title{Get the large vector from sono}
\usage{
get_sono_large(sono)
}
\arguments{
\item{sono}{Sonogram vector in Xeno Canto record}
}
\value{
sono$large
}
\description{
Get the large vector from sono
}
