% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_me.R
\name{multiple_choice}
\alias{multiple_choice}
\title{Test type: multiple choice}
\usage{
multiple_choice(secret_record)
}
\arguments{
\item{secret_record}{a row from records_df}
}
\value{
invisible
}
\description{
Test type: multiple choice
}
\examples{
\dontrun{multiple_choice(records_df[1])}
}
