% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_recording.R
\name{play_recording}
\alias{play_recording}
\title{Play a recording from xeno canto}
\usage{
play_recording(x)
}
\arguments{
\item{x}{path to a MP3 file}
}
\value{
nothing
}
\description{
Play a recording from xeno canto
}
