% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_xeno_canto.R
\name{req_xeno}
\alias{req_xeno}
\title{Build a request to Xeno Canto for a single species}
\usage{
req_xeno(species_name = "troglodytes+troglodytes")
}
\arguments{
\item{species_name}{Character. The name of the species to send to Xeno Canto. See details for more. "troglodytes+troglodytes"}
}
\value{
A modified HTTP request
}
\description{
Build a request to Xeno Canto for a single species
}
\details{
species_name can be any character string accepted by the Xeno Canto API.
See https://xeno-canto.org/help/search for details.
Known options include:
\itemize{
\item Genus species "troglodytes+troglodytes" or "troglodytes troglodytes"
}
}
\examples{
req_xeno()
}
