% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_xeno_canto.R
\name{req_xeno_perform}
\alias{req_xeno_perform}
\title{Perform a request to Xeno Canto}
\usage{
req_xeno_perform(species_name = "troglodytes+troglodytes")
}
\arguments{
\item{species_name}{Character. The name of the species to send to Xeno Canto. See details for more. "troglodytes+troglodytes"}
}
\value{
The json body of an http request
}
\description{
Perform a request to Xeno Canto
}
\details{
species_name can be any character string accepted by the Xeno Canto API.
See https://xeno-canto.org/help/search for details.
Known options include:
\itemize{
\item Genus species "troglodytes+troglodytes" or "troglodytes troglodytes"
}

NB that it will wait for a short time (between 1.1 and 3 seconds) between
requests so as not to overwhelm Xeno Canto
}
\examples{
req_xeno_perform()
}
