% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_xeno_canto.R
\name{search_xeno_canto}
\alias{search_xeno_canto}
\title{Perform multiple requests to Xeno Canto}
\usage{
search_xeno_canto(
  species_vect = c("troglodytes+troglodytes", "icterus+abeillei")
)
}
\arguments{
\item{species_vect}{A character vector of species to request from Xeno Canto}
}
\value{
A list of http responses from Xeno Canto
}
\description{
Perform multiple requests to Xeno Canto
}
\examples{
\dontrun{
search_xeno_canto("troglodytes+troglodytes")
}
}
