% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_me.R
\name{test_me}
\alias{test_me}
\title{Test your song knowledge}
\usage{
test_me(
  records_df,
  test_type = c(true_false, multiple_choice, free_choice, flashcard),
  ...
)
}
\arguments{
\item{records_df}{dataframe from download_recordings}

\item{test_type}{function describing how you would like to be tested. See details}

\item{...}{parameters passed to the test_type functions}
}
\value{
nothing
}
\description{
Test your song knowledge
}
\examples{
\dontrun{
spp_lst <- collect_species_list("Vancouver, BC")
resps <- search_xeno_canto(spp_lst)
recs <- download_recordings(resps, save_directory = here::here())
test_me(recs)
}
}
