% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_me.R
\name{true_false}
\alias{true_false}
\title{Test type: true or false}
\usage{
true_false(secret_record)
}
\arguments{
\item{secret_record}{a row from records_df}
}
\value{
invisible
}
\description{
Test type: true or false
}
\examples{
\dontrun{true_false(records_df[1])}
}
